Refactor InkCanvas drawing architecture
