+++ sources/Experiments/InkCanvas/App.cs
@@ -1,10 +1,10 @@
 ﻿using System.Numerics;
+using InkCanvas.Drawing;
 using InkCanvas.Helpers;
 using Silk.NET.Input;
 using Silk.NET.Windowing;
 using Zenith.NET;
 using Zenith.NET.DirectX12;
-using Zenith.NET.Extensions.Skia;
 using Zenith.NET.Metal;
 using Zenith.NET.Vulkan;
 
@@ -15,9 +15,7 @@ internal static class App
     private static readonly IWindow window;
     private static readonly IInputContext input;
     private static readonly SwapChain swapChain;
-    private static readonly Board board;
-
-    private static SKTexture texture;
+    private static readonly CanvasController canvas;
 
     static App()
     {
@@ -72,15 +70,7 @@ internal static class App
             Format = PixelFormat.B8G8R8A8UNorm
         });
 
-        board = new();
-
-        texture = Context.CreateSKTexture(new()
-        {
-            Format = PixelFormat.B8G8R8A8UNorm,
-            Width = Width,
-            Height = Height,
-            SampleCount = SampleCount.Count1
-        });
+        canvas = new(Context, input, Width, Height);
     }
 
     public static GraphicsContext Context { get; }
@@ -100,32 +90,10 @@ internal static class App
                 return;
             }
 
-            uint width = (uint)(Width / DpiScale.X);
-            uint height = (uint)(Height / DpiScale.Y);
-
-            texture.Render((canvas) =>
-            {
-                canvas.Save();
-                canvas.Scale(DpiScale.X, DpiScale.Y);
-
-                board.Draw(canvas, Width / DpiScale.X, Height / DpiScale.Y);
-
-                canvas.Restore();
-            });
-
             CommandBuffer commandBuffer = Context.GraphicsQueue.CommandBuffer();
 
             commandBuffer.Transition(swapChain.Drawable, default, TextureLayout.Undefined, TextureLayout.CopyDst);
-            commandBuffer.Transition(texture, default, TextureLayout.ColorAttachment, TextureLayout.CopySrc);
-
-            commandBuffer.CopyTexture(texture, default, default, swapChain.Drawable, default, default, new()
-            {
-                Width = width,
-                Height = height,
-                Depth = 1
-            });
-
-            commandBuffer.Transition(texture, default, TextureLayout.CopySrc, TextureLayout.ColorAttachment);
+            canvas.Render(commandBuffer, swapChain.Drawable, DpiScale);
             commandBuffer.Transition(swapChain.Drawable, default, TextureLayout.CopyDst, TextureLayout.Present);
 
             commandBuffer.Submit().Wait();
@@ -133,48 +101,20 @@ internal static class App
             swapChain.Present();
         };
 
-        window.Resize += _ =>
+        window.FramebufferResize += _ =>
         {
             if (Width is 0 || Height is 0)
             {
                 return;
             }
 
-            texture.Dispose();
-            texture = Context.CreateSKTexture(new()
-            {
-                Format = PixelFormat.B8G8R8A8UNorm,
-                Width = Width,
-                Height = Height,
-                SampleCount = SampleCount.Count1
-            });
-
+            canvas.Resize(Width, Height);
             swapChain.Resize(Width, Height);
-        };
-
-        IMouse mouse = input.Mice[0];
-        mouse.MouseMove += (_, position) => board.PointerMove(new(position.X, position.Y));
-
-        mouse.MouseDown += (_, button) =>
-        {
-            if (button is MouseButton.Left or MouseButton.Right)
-            {
-                board.PointerDown(new(mouse.Position.X, mouse.Position.Y), button is MouseButton.Right);
-            }
-        };
-
-        mouse.MouseUp += (_, button) =>
-        {
-            if (button is MouseButton.Left or MouseButton.Right)
-            {
-                board.PointerUp(button is MouseButton.Right);
-            }
         };
 
         window.Run();
 
-        board.Dispose();
-        texture.Dispose();
+        canvas.Dispose();
         swapChain.Dispose();
         input.Dispose();
         window.Dispose();
